Extract base class for shared functionality
